refactor(storybook): move custom addons out of src
