+++ .storybook/webpack.config.js
@@ -1,7 +1,18 @@
+const path = require('path');
 const defaultConfig = require('@stoplight/scripts/.storybook/webpack.config');
 
 module.exports = (baseConfig, env, config) => {
   config = defaultConfig(baseConfig, env, config);
+
+  config.module.rules.push({
+    test: /\.(ts|tsx)$/,
+    include: [path.resolve(process.cwd(), '.storybook')],
+    use: [
+      {
+        loader: require.resolve('ts-loader'),
+      },
+    ],
+  });
 
   // ... further customize if needed
